google-storage-bucket-source/requirements.txt
quixstreams[s3]==3.22.0
python-dotenv
google-cloud-storage
quixstreams
